pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-silk>=5.4.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.1",
    "pydot>=4.0.1",
    "uuid6>=2025.0.1",
]
